requirements_api.txt
# FastAPI and web server dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19

# Your existing dependencies (core RAG system)
langchain==0.3.9
langchain-community==0.3.8
langchain-pinecone==0.2.0
langchain-groq==0.2.1
langchain-huggingface==0.3.1
sentence-transformers==5.1.1
pinecone-client==5.0.1
groq==0.13.0
python-dotenv==1.1.1
torch==2.8.0
transformers==4.56.2

# LlamaParse (if using)
llama-cloud-services